requirements.txt
torch
polars
mlflow
ruff
hydra-core